requirements.txt
beautifulsoup4==4.4.0
cassandra-driver==2.7.1
click==5.1
decorator==4.0.2
futures==3.0.3
GDAL==2.0.0
ipython==4.0.0
ipython-genutils==0.1.0
lxml==3.4.4
numpy==1.9.2
path.py==0.0.0
pexpect==3.3
pickleshare==0.5
python-dateutil==2.4.2
python-dotenv==0.1.3
pytz==2015.4
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
wheel==0.24.0
